Add files via upload
One of four homework assignments for the term. The assignment consisted of an in-depth look at Google Analytics featuring the specific case of the Whitman School of Business's campaign for their online MBA. Simulated evaluating trends for market campaign improvement and evaluating traffic flow to recommend sources of user interface inadequacy.