Add files via upload
Data Exploration and Linear Model project looking at NBA data. The classification problem of design was trying to pick characteristics of players from the data that were most indicative of playing in the Championship finals or being selected the regular season MVP. This project was definitely more ambitious in scope but much was learned in completing this project.