Add files via upload
This final homework assignment of the course focused on topic classification for the minutes of speeches made by representatives in the house. Comparing topics to party and gender affiliation is important for all people to know so that they can better empathize and understand other viewpoints. The assignment used Latent-Direlich Allocation to pick prominent keywords that differentiated all of the speeches. A succinct title was then created that best fit the top keywords. It should be noted that the topic titles were ascribed manually and thus are open to argument as to whether the title effectively matches the top keywords.